+++ site/dat/docs/Changelog.md
@@ -10,6 +10,7 @@
  - add `action` blocks to JMeter's requests
  - fix percentiles values for cloud-based tests
  - support collection-based cloud tests
+ - add stubs for `screenshoter` and `capturehar` services
  - add RSpec tests runner for Selenium
 
 ## 1.6.8 <sup>1 sep 2016</sup>